pyproject.toml
[tool.poetry]
name = "distribd"
version = "0.1.0"
description = "A simple replicated docker image registry"
authors = ["John Carr <john.carr@unrouted.co.uk>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.6.2"
aiofile = "^1.5.2"
coloredlogs = "^14.0"
verboselogs = "^1.7"
prometheus_client = "^0.7.1"
aiojobs = "^0.2.2"
jsonschema = "^3.2.0"
yarl = "^1.4.2"
pyjwt = {version = "^1.7.1", extras = ["crypto"]}
confuse = "^1.1.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
pytest = "^5.4.1"
isort = "^4.3.21"
pytest-aiohttp = "^0.3.0"
black = "^19.10b0"
flake8 = "^3.7.9"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
